Struct google_androidpublisher2::EditMethods

pub struct EditMethods<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on edit resources. It is not used directly, but through the AndroidPublisher hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_androidpublisher2 as androidpublisher2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use androidpublisher2::AndroidPublisher;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AndroidPublisher::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `apklistings_delete(...)`, `apklistings_deleteall(...)`, `apklistings_get(...)`, `apklistings_list(...)`, `apklistings_patch(...)`, `apklistings_update(...)`, `apks_addexternallyhosted(...)`, `apks_list(...)`, `apks_upload(...)`, `bundles_list(...)`, `bundles_upload(...)`, `commit(...)`, `delete(...)`, `deobfuscationfiles_upload(...)`, `details_get(...)`, `details_patch(...)`, `details_update(...)`, `expansionfiles_get(...)`, `expansionfiles_patch(...)`, `expansionfiles_update(...)`, `expansionfiles_upload(...)`, `get(...)`, `images_delete(...)`, `images_deleteall(...)`, `images_list(...)`, `images_upload(...)`, `insert(...)`, `listings_delete(...)`, `listings_deleteall(...)`, `listings_get(...)`, `listings_list(...)`, `listings_patch(...)`, `listings_update(...)`, `testers_get(...)`, `testers_patch(...)`, `testers_update(...)`, `tracks_get(...)`, `tracks_list(...)`, `tracks_patch(...)`, `tracks_update(...)` and `validate(...)`
// to build up your call.
let rb = hub.edits();

Methods

impl<'a, C, A> EditMethods<'a, C, A>

pub fn deobfuscationfiles_upload(
&self,
package_name: &str,
edit_id: &str,
apk_version_code: i32,
deobfuscation_file_type: &str
) -> EditDeobfuscationfileUploadCall<'a, C, A>

Create a builder to help you perform the following task:

Uploads the deobfuscation file of the specified APK. If a deobfuscation file already exists, it will be replaced.

Arguments

packageName - Unique identifier of the Android app for which the deobfuscatiuon files are being uploaded; for example, "com.spiffygame".

editId - Unique identifier for this edit.

apkVersionCode - The version code of the APK whose deobfuscation file is being uploaded.

deobfuscationFileType - No description provided.

pub fn images_upload(
&self,
package_name: &str,
edit_id: &str,
language: &str,
image_type: &str
) -> EditImageUploadCall<'a, C, A>

Create a builder to help you perform the following task:

Uploads a new image and adds it to the list of images for the specified language and image type.

Arguments

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

language - The language code (a BCP-47 language tag) of the localized listing whose images are to read or modified. For example, to select Austrian German, pass "de-AT".

imageType - No description provided.

pub fn expansionfiles_update(
&self,
request: ExpansionFile,
package_name: &str,
edit_id: &str,
apk_version_code: i32,
expansion_file_type: &str
) -> EditExpansionfileUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the APK's Expansion File configuration to reference another APK's Expansion Files. To add a new Expansion File use the Upload method.

Arguments

request - No description provided.

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

apkVersionCode - The version code of the APK whose Expansion File configuration is being read or modified.

expansionFileType - No description provided.

pub fn details_get(
&self,
package_name: &str,
edit_id: &str
) -> EditDetailGetCall<'a, C, A>

Create a builder to help you perform the following task:

Fetches app details for this edit. This includes the default language and developer support contact information.

Arguments

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

pub fn apklistings_delete(
&self,
package_name: &str,
edit_id: &str,
apk_version_code: i32,
language: &str
) -> EditApklistingDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the APK-specific localized listing for a specified APK and language code.

Arguments

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

apkVersionCode - The APK version code whose APK-specific listings should be read or modified.

language - The language code (a BCP-47 language tag) of the APK-specific localized listing to read or modify. For example, to select Austrian German, pass "de-AT".

pub fn listings_deleteall(
&self,
package_name: &str,
edit_id: &str
) -> EditListingDeleteallCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes all localized listings from an edit.

Arguments

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

pub fn apks_addexternallyhosted(
&self,
request: ApksAddExternallyHostedRequest,
package_name: &str,
edit_id: &str
) -> EditApkAddexternallyhostedCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new APK without uploading the APK itself to Google Play, instead hosting the APK at a specified URL. This function is only available to enterprises using Google Play for Work whose application is configured to restrict distribution to the enterprise domain.

Arguments

request - No description provided.

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

pub fn apklistings_deleteall(
&self,
package_name: &str,
edit_id: &str,
apk_version_code: i32
) -> EditApklistingDeleteallCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes all the APK-specific localized listings for a specified APK.

Arguments

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

apkVersionCode - The APK version code whose APK-specific listings should be read or modified.

pub fn details_update(
&self,
request: AppDetails,
package_name: &str,
edit_id: &str
) -> EditDetailUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates app details for this edit.

Arguments

request - No description provided.

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

pub fn tracks_get(
&self,
package_name: &str,
edit_id: &str,
track: &str
) -> EditTrackGetCall<'a, C, A>

Create a builder to help you perform the following task:

Fetches the track configuration for the specified track type. Includes the APK version codes that are in this track.

Arguments

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

track - The track to read or modify.

pub fn expansionfiles_patch(
&self,
request: ExpansionFile,
package_name: &str,
edit_id: &str,
apk_version_code: i32,
expansion_file_type: &str
) -> EditExpansionfilePatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the APK's Expansion File configuration to reference another APK's Expansion Files. To add a new Expansion File use the Upload method. This method supports patch semantics.

Arguments

request - No description provided.

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

apkVersionCode - The version code of the APK whose Expansion File configuration is being read or modified.

expansionFileType - No description provided.

pub fn images_list(
&self,
package_name: &str,
edit_id: &str,
language: &str,
image_type: &str
) -> EditImageListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all images for the specified language and image type.

Arguments

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

language - The language code (a BCP-47 language tag) of the localized listing whose images are to read or modified. For example, to select Austrian German, pass "de-AT".

imageType - No description provided.

pub fn tracks_update(
&self,
request: Track,
package_name: &str,
edit_id: &str,
track: &str
) -> EditTrackUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the track configuration for the specified track type.

Arguments

request - No description provided.

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

track - The track to read or modify.

pub fn listings_patch(
&self,
request: Listing,
package_name: &str,
edit_id: &str,
language: &str
) -> EditListingPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Creates or updates a localized store listing. This method supports patch semantics.

Arguments

request - No description provided.

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

language - The language code (a BCP-47 language tag) of the localized listing to read or modify. For example, to select Austrian German, pass "de-AT".

pub fn get(&self, package_name: &str, edit_id: &str) -> EditGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns information about the edit specified. Calls will fail if the edit is no long active (e.g. has been deleted, superseded or expired).

Arguments

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

pub fn images_delete(
&self,
package_name: &str,
edit_id: &str,
language: &str,
image_type: &str,
image_id: &str
) -> EditImageDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the image (specified by id) from the edit.

Arguments

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

language - The language code (a BCP-47 language tag) of the localized listing whose images are to read or modified. For example, to select Austrian German, pass "de-AT".

imageType - No description provided.

imageId - Unique identifier an image within the set of images attached to this edit.

pub fn apks_upload(
&self,
package_name: &str,
edit_id: &str
) -> EditApkUploadCall<'a, C, A>

Arguments

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

pub fn apklistings_update(
&self,
request: ApkListing,
package_name: &str,
edit_id: &str,
apk_version_code: i32,
language: &str
) -> EditApklistingUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates or creates the APK-specific localized listing for a specified APK and language code.

Arguments

request - No description provided.

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

apkVersionCode - The APK version code whose APK-specific listings should be read or modified.

language - The language code (a BCP-47 language tag) of the APK-specific localized listing to read or modify. For example, to select Austrian German, pass "de-AT".

pub fn apks_list(
&self,
package_name: &str,
edit_id: &str
) -> EditApkListCall<'a, C, A>

Arguments

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

pub fn apklistings_patch(
&self,
request: ApkListing,
package_name: &str,
edit_id: &str,
apk_version_code: i32,
language: &str
) -> EditApklistingPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates or creates the APK-specific localized listing for a specified APK and language code. This method supports patch semantics.

Arguments

request - No description provided.

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

apkVersionCode - The APK version code whose APK-specific listings should be read or modified.

language - The language code (a BCP-47 language tag) of the APK-specific localized listing to read or modify. For example, to select Austrian German, pass "de-AT".

pub fn listings_get(
&self,
package_name: &str,
edit_id: &str,
language: &str
) -> EditListingGetCall<'a, C, A>

Create a builder to help you perform the following task:

Fetches information about a localized store listing.

Arguments

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

language - The language code (a BCP-47 language tag) of the localized listing to read or modify. For example, to select Austrian German, pass "de-AT".

pub fn bundles_list(
&self,
package_name: &str,
edit_id: &str
) -> EditBundleListCall<'a, C, A>

Arguments

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

pub fn delete(
&self,
package_name: &str,
edit_id: &str
) -> EditDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes an edit for an app. Creating a new edit will automatically delete any of your previous edits so this method need only be called if you want to preemptively abandon an edit.

Arguments

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

pub fn expansionfiles_upload(
&self,
package_name: &str,
edit_id: &str,
apk_version_code: i32,
expansion_file_type: &str
) -> EditExpansionfileUploadCall<'a, C, A>

Create a builder to help you perform the following task:

Uploads and attaches a new Expansion File to the APK specified.

Arguments

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

apkVersionCode - The version code of the APK whose Expansion File configuration is being read or modified.

expansionFileType - No description provided.

pub fn insert(
&self,
request: AppEdit,
package_name: &str
) -> EditInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new edit for an app, populated with the app's current state.

Arguments

request - No description provided.

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

pub fn listings_list(
&self,
package_name: &str,
edit_id: &str
) -> EditListingListCall<'a, C, A>

Create a builder to help you perform the following task:

Returns all of the localized store listings attached to this edit.

Arguments

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

pub fn testers_patch(
&self,
request: Testers,
package_name: &str,
edit_id: &str,
track: &str
) -> EditTesterPatchCall<'a, C, A>

Arguments

request - No description provided.

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

track - The track to read or modify.

pub fn commit(
&self,
package_name: &str,
edit_id: &str
) -> EditCommitCall<'a, C, A>

Create a builder to help you perform the following task:

Commits/applies the changes made in this edit back to the app.

Arguments

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

pub fn tracks_list(
&self,
package_name: &str,
edit_id: &str
) -> EditTrackListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all the track configurations for this edit.

Arguments

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

pub fn testers_get(
&self,
package_name: &str,
edit_id: &str,
track: &str
) -> EditTesterGetCall<'a, C, A>

Arguments

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

track - The track to read or modify.

pub fn validate(
&self,
package_name: &str,
edit_id: &str
) -> EditValidateCall<'a, C, A>

Create a builder to help you perform the following task:

Checks that the edit can be successfully committed. The edit's changes are not applied to the live app.

Arguments

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

pub fn bundles_upload(
&self,
package_name: &str,
edit_id: &str
) -> EditBundleUploadCall<'a, C, A>

Create a builder to help you perform the following task:

Uploads a new Android App Bundle to this edit. If you are using the Google API client libraries, please increase the timeout of the http request before calling this endpoint (a timeout of 2 minutes is recommended). See: https://developers.google.com/api-client-library/java/google-api-java-client/errors for an example in java.

Arguments

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

pub fn listings_update(
&self,
request: Listing,
package_name: &str,
edit_id: &str,
language: &str
) -> EditListingUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates or updates a localized store listing.

Arguments

request - No description provided.

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

language - The language code (a BCP-47 language tag) of the localized listing to read or modify. For example, to select Austrian German, pass "de-AT".

pub fn expansionfiles_get(
&self,
package_name: &str,
edit_id: &str,
apk_version_code: i32,
expansion_file_type: &str
) -> EditExpansionfileGetCall<'a, C, A>

Create a builder to help you perform the following task:

Fetches the Expansion File configuration for the APK specified.

Arguments

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

apkVersionCode - The version code of the APK whose Expansion File configuration is being read or modified.

expansionFileType - No description provided.

pub fn images_deleteall(
&self,
package_name: &str,
edit_id: &str,
language: &str,
image_type: &str
) -> EditImageDeleteallCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes all images for the specified language and image type.

Arguments

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

language - The language code (a BCP-47 language tag) of the localized listing whose images are to read or modified. For example, to select Austrian German, pass "de-AT".

imageType - No description provided.

pub fn details_patch(
&self,
request: AppDetails,
package_name: &str,
edit_id: &str
) -> EditDetailPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates app details for this edit. This method supports patch semantics.

Arguments

request - No description provided.

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

pub fn tracks_patch(
&self,
request: Track,
package_name: &str,
edit_id: &str,
track: &str
) -> EditTrackPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the track configuration for the specified track type. This method supports patch semantics.

Arguments

request - No description provided.

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

track - The track to read or modify.

pub fn listings_delete(
&self,
package_name: &str,
edit_id: &str,
language: &str
) -> EditListingDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified localized store listing from an edit.

Arguments

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

language - The language code (a BCP-47 language tag) of the localized listing to read or modify. For example, to select Austrian German, pass "de-AT".

pub fn apklistings_get(
&self,
package_name: &str,
edit_id: &str,
apk_version_code: i32,
language: &str
) -> EditApklistingGetCall<'a, C, A>

Create a builder to help you perform the following task:

Fetches the APK-specific localized listing for a specified APK and language code.

Arguments

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

apkVersionCode - The APK version code whose APK-specific listings should be read or modified.

language - The language code (a BCP-47 language tag) of the APK-specific localized listing to read or modify. For example, to select Austrian German, pass "de-AT".

pub fn testers_update(
&self,
request: Testers,
package_name: &str,
edit_id: &str,
track: &str
) -> EditTesterUpdateCall<'a, C, A>

Arguments

request - No description provided.

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

track - The track to read or modify.

pub fn apklistings_list(
&self,
package_name: &str,
edit_id: &str,
apk_version_code: i32
) -> EditApklistingListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all the APK-specific localized listings for a specified APK.

Arguments

packageName - Unique identifier for the Android app that is being updated; for example, "com.spiffygame".

editId - Unique identifier for this edit.

apkVersionCode - The APK version code whose APK-specific listings should be read or modified.