Struct google_androidpublisher2::EntitlementMethods

pub struct EntitlementMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on entitlement resources. It is not used directly, but through the AndroidPublisher hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_androidpublisher2 as androidpublisher2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use androidpublisher2::AndroidPublisher;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AndroidPublisher::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `list(...)`
// to build up your call.
let rb = hub.entitlements();

Methods

impl<'a, C, A> EntitlementMethods<'a, C, A>

fn list(&self, package_name: &str) -> EntitlementListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the user's current inapp item or subscription entitlements

Arguments

packageName - The package name of the application the inapp product was sold in (for example, 'com.some.thing').