Struct google_androidpublisher2::InAppProduct

pub struct InAppProduct {
    pub sku: Option<String>,
    pub status: Option<String>,
    pub subscription_period: Option<String>,
    pub season: Option<Season>,
    pub package_name: Option<String>,
    pub listings: Option<HashMap<String, InAppProductListing>>,
    pub trial_period: Option<String>,
    pub purchase_type: Option<String>,
    pub default_language: Option<String>,
    pub prices: Option<HashMap<String, Price>>,
    pub default_price: Option<Price>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

patch inappproducts (request|response)

insert inappproducts (request|response)

get inappproducts (response)

update inappproducts (request|response)

Fields

sku

The stock-keeping-unit (SKU) of the product, unique within an app.

status

no description provided

subscription_period

The period of the subscription (if any), i.e. period at which payments must happen. Defined as ISO 8601 duration, i.e. "P1M" for 1 month period.

season

Definition of a season for a seasonal subscription. Can be defined only for yearly subscriptions.

package_name

The package name of the parent app.

listings

List of localized title and description data.

trial_period

Trial period, specified in ISO 8601 format. Acceptable values are anything between "P7D" (seven days) and "P999D" (999 days). Seasonal subscriptions cannot have a trial period.

purchase_type

Purchase type enum value. Unmodifiable after creation.

default_language

The default language of the localized data, as defined by BCP 47. e.g. "en-US", "en-GB".

prices

Prices per buyer region. None of these prices should be zero. In-app products can never be free.

default_price

Default price cannot be zero. In-app products can never be free. Default price is always in the developer's Checkout merchant currency.