Struct google_androidpublisher2::InappproductMethods

pub struct InappproductMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on inappproduct resources. It is not used directly, but through the AndroidPublisher hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_androidpublisher2 as androidpublisher2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use androidpublisher2::AndroidPublisher;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AndroidPublisher::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `batch(...)`, `delete(...)`, `get(...)`, `insert(...)`, `list(...)`, `patch(...)` and `update(...)`
// to build up your call.
let rb = hub.inappproducts();

Methods

impl<'a, C, A> InappproductMethods<'a, C, A>

fn batch(&self, request: InappproductsBatchRequest) -> InappproductBatchCall<'a, C, A>

Arguments

request - No description provided.

fn list(&self, package_name: &str) -> InappproductListCall<'a, C, A>

Create a builder to help you perform the following task:

List all the in-app products for an Android app, both subscriptions and managed in-app products..

Arguments

packageName - Unique identifier for the Android app with in-app products; for example, "com.spiffygame".

fn insert(&self, request: InAppProduct, package_name: &str) -> InappproductInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new in-app product for an app.

Arguments

request - No description provided.

packageName - Unique identifier for the Android app; for example, "com.spiffygame".

fn delete(&self, package_name: &str, sku: &str) -> InappproductDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Delete an in-app product for an app.

Arguments

packageName - Unique identifier for the Android app with the in-app product; for example, "com.spiffygame".

sku - Unique identifier for the in-app product.

fn get(&self, package_name: &str, sku: &str) -> InappproductGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns information about the in-app product specified.

Arguments

packageName - No description provided.

sku - Unique identifier for the in-app product.

fn update(&self, request: InAppProduct, package_name: &str, sku: &str) -> InappproductUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the details of an in-app product.

Arguments

request - No description provided.

packageName - Unique identifier for the Android app with the in-app product; for example, "com.spiffygame".

sku - Unique identifier for the in-app product.

fn patch(&self, request: InAppProduct, package_name: &str, sku: &str) -> InappproductPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the details of an in-app product. This method supports patch semantics.

Arguments

request - No description provided.

packageName - Unique identifier for the Android app with the in-app product; for example, "com.spiffygame".

sku - Unique identifier for the in-app product.