Struct google-androidpublisher2::PurchaseMethodsBuilder

pub struct PurchaseMethodsBuilder<'a, C, NC, A> where C: 'a, NC: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on purchase resources. It is not used directly, but through the AndroidPublisher hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate serde;
extern crate "google-androidpublisher2" as androidpublisher2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use androidpublisher2::AndroidPublisher;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AndroidPublisher::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `products_get(...)`, `subscriptions_cancel(...)`, `subscriptions_defer(...)`, `subscriptions_get(...)`, `subscriptions_refund(...)` and `subscriptions_revoke(...)`
// to build up your call.
let rb = hub.purchases();

Methods

impl<'a, C, NC, A> PurchaseMethodsBuilder<'a, C, NC, A>

fn subscriptions_revoke(&self, package_name: &str, subscription_id: &str, token: &str) -> PurchaseSubscriptionRevokeCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Refunds and immediately revokes a user's subscription purchase. Access to the subscription will be terminated immediately and it will stop recurring.

fn products_get(&self, package_name: &str, product_id: &str, token: &str) -> PurchaseProductGetCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Checks the purchase and consumption status of an inapp item.

fn subscriptions_get(&self, package_name: &str, subscription_id: &str, token: &str) -> PurchaseSubscriptionGetCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Checks whether a user's subscription purchase is valid and returns its expiry time.

fn subscriptions_cancel(&self, package_name: &str, subscription_id: &str, token: &str) -> PurchaseSubscriptionCancelCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Cancels a user's subscription purchase. The subscription remains valid until its expiration time.

fn subscriptions_refund(&self, package_name: &str, subscription_id: &str, token: &str) -> PurchaseSubscriptionRefundCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Refunds a user's subscription purchase, but the subscription remains valid until its expiration time and it will continue to recur.

fn subscriptions_defer(&self, request: &SubscriptionPurchasesDeferRequest, package_name: &str, subscription_id: &str, token: &str) -> PurchaseSubscriptionDeferCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Defers a user's subscription purchase until a specified future expiration time.