Struct google_androidpublisher2::ExternallyHostedApk

pub struct ExternallyHostedApk {
    pub icon_base64: Option<String>,
    pub certificate_base64s: Option<Vec<String>>,
    pub externally_hosted_url: Option<String>,
    pub maximum_sdk: Option<i32>,
    pub file_sha256_base64: Option<String>,
    pub uses_permissions: Option<Vec<ExternallyHostedApkUsesPermission>>,
    pub file_sha1_base64: Option<String>,
    pub uses_features: Option<Vec<String>>,
    pub native_codes: Option<Vec<String>>,
    pub file_size: Option<String>,
    pub version_name: Option<String>,
    pub version_code: Option<i32>,
    pub minimum_sdk: Option<i32>,
    pub application_label: Option<String>,
    pub package_name: Option<String>,
}

Defines an APK available for this application that is hosted externally and not uploaded to Google Play. This function is only available to enterprises who are using Google Play for Work, and whos application is restricted to the enterprise private channel

This type is not used in any activity, and only used as part of another schema.

Fields

icon_base64

The icon image from the APK, as a base64 encoded byte array.

certificate_base64s

A certificate (or array of certificates if a certificate-chain is used) used to signed this APK, represented as a base64 encoded byte array.

externally_hosted_url

The URL at which the APK is hosted. This must be an https URL.

maximum_sdk

The maximum SDK supported by this APK (optional).

file_sha256_base64

The SHA256 checksum of this APK, represented as a base64 encoded byte array.

uses_permissions

The permissions requested by this APK.

file_sha1_base64

The SHA1 checksum of this APK, represented as a base64 encoded byte array.

uses_features

The features required by this APK (optional).

native_codes

The native code environments supported by this APK (optional).

file_size

The file size in bytes of this APK.

version_name

The version name of this APK.

version_code

The version code of this APK.

minimum_sdk

The minimum SDK targeted by this APK.

application_label

The application label.

package_name

The package name.