Struct google_androidpublisher2::SubscriptionPurchase

pub struct SubscriptionPurchase {
    pub kind: Option<String>,
    pub country_code: Option<String>,
    pub developer_payload: Option<String>,
    pub auto_renewing: Option<bool>,
    pub payment_state: Option<i32>,
    pub price_currency_code: Option<String>,
    pub expiry_time_millis: Option<String>,
    pub cancel_reason: Option<i32>,
    pub start_time_millis: Option<String>,
    pub price_amount_micros: Option<String>,
}

A SubscriptionPurchase resource indicates the status of a user's subscription purchase.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

subscriptions get purchases (response)

Fields
kind: Option<String>

This kind represents a subscriptionPurchase object in the androidpublisher service.

country_code: Option<String>

ISO 3166-1 alpha-2 billing country/region code of the user at the time the subscription was granted.

developer_payload: Option<String>

A developer-specified string that contains supplemental information about an order.

auto_renewing: Option<bool>

Whether the subscription will automatically be renewed when it reaches its current expiry time.

payment_state: Option<i32>

The payment state of the subscription. Possible values are: - Payment pending - Payment received

price_currency_code: Option<String>

ISO 4217 currency code for the subscription price. For example, if the price is specified in British pounds sterling, price_currency_code is "GBP".

expiry_time_millis: Option<String>

Time at which the subscription will expire, in milliseconds since the Epoch.

cancel_reason: Option<i32>

The reason why a subscription was cancelled or is not auto-renewing. Possible values are: - User cancelled the subscription - Subscription was cancelled by the system, for example because of a billing problem

start_time_millis: Option<String>

Time at which the subscription was granted, in milliseconds since the Epoch.

price_amount_micros: Option<String>

Price of the subscription, not including tax. Price is expressed in micro-units, where 1,000,000 micro-units represents one unit of the currency. For example, if the subscription price is €1.99, price_amount_micros is 1990000.